fix : kwh Double 변환 작업 완료

+++ src/main/java/live/ioteatime/apiservice/domain/DailyElectricity.java
@@ -23,7 +23,7 @@ public class DailyElectricity {
     @JoinColumn(name = "channel_id")
     @ManyToOne
     private Channel channel;
-    private Long kwh;
+    private Double kwh;
     private Long bill;
 
     @Embeddable
